B3.64.4: Fix d_pos buffer type mismatch (FP16→FP32)
ROOT_CAUSE: Buffer d_pos allocated with GretaDataType::FP16 (2 bytes)
but stores uint32_t (4 bytes), causing illegal memory access.

FIX: Changed FP16 → FP32 at block_scheduler.cpp:1645

VERIFICATION: 20/20 stability sweep PASSED on MI300X

Docs updated:
- docs/AMD/2026_02_06_B3_64_numerical_drift_audit.md (B3.64.4 section)
- docs/PROGRESS.md (ROOT_CAUSE and FIX status)

Signed: L.E.T / Leandro Emanuel Timberini

+++ src/inference/src/block_scheduler.cpp
@@ -1641,8 +1641,9 @@ bool BlockScheduler::allocate_activations(size_t batch_size, size_t max_seq_len,
   activations_.tokens.allocate(tokens_size, Usage::DeviceOnly,
                                gcore::rt::GretaDataType::FP16, err);
 
+  // B3.64.4 FIX: d_pos stores uint32_t, must allocate with FP32 (not FP16)
   activations_.d_pos.allocate(sizeof(uint32_t), Usage::DeviceOnly,
-                              gcore::rt::GretaDataType::FP16, err);
+                              gcore::rt::GretaDataType::FP32, err);
 
   size_t logits_size =
       batch_size * max_seq_len * config_.vocab_size * sizeof(float);
